### 2017-06-26 #306 @whedon
[REVIEW]: PyGBe-LSPR: Python and GPU Boundary-integral solver for electrostatics
**Submitting author:** @ncclementi (<a href="http://orcid.org/0000-0002-0575-5520">Natalia C. Clementi</a>)
**Repository:** <a href="https://github.com/barbagroup/pygbe" target ="_blank">https://github.com/barbagroup/pygbe</a>
**Version:** 0.3
