### 2017-07-01 #306 @labarba
@nicoguaro : Do you have access to a GPU to run this? Let us know if you need anything.
