### 2017-07-10 #306 @nicoguaro
@labarba, I have a GTX 850M in my laptop... but it just stopped working. I will try in a computer in the office that has a G98 Quadro NVS 420.

It seems a little bit more complicated than I thought.
