### 2017-07-14 #306 @kyleniemeyer
@katyhuff I can jump in to review this if needed, as I do have access to a few GPU systems (and reviewed the original PyGBe submission)
