### 2017-07-14 #306 @katyhuff
Thanks a lot @nicoguaro for giving this a look! The particular hardware constraints of this paper may require an editor with access to an appropriate GPU resource. In the future, we may have a journal-wide solution to unique architectural requirements. After some discussion with him here at SciPy, I believe that @kyleniemeyer may be able to conduct the review with the GPU resources he has available to him. @kyleniemeyer , would you be willing to review this paper?
