### 2017-08-17 barbagroup/pygbe#56 @kyleniemeyer
Error running tests
Hi there, I'm working my review for JOSS (https://github.com/openjournals/joss-reviews/issues/306), but running into errors when I try to run the test suite.

I followed your conda installation instructions, and my PyCUDA install seems to be ok (tests pass; though I did install version `2017.1`). However, when I try to run `pytest -v` from the top-level `pygbe` directory, I get errors on multiple tests that look like:
```
>       with open('lspr.pickle', 'rb') as f:
E       FileNotFoundError: [Errno 2] No such file or directory: 'lspr.pickle'
```

Otherwise I was (so far) able to run the `pygbe examples/lys` example, so my build/install seems fine.

Any ideas?
