### 2017-08-17 barbagroup/pygbe#56 @kyleniemeyer
I think I know what may fix this—I think you need to add `include pygbe/tests/*.pickle` to your `MANIFEST.in` file to ensure the non `.py` files get copied in the install.
