### 2017-08-17 barbagroup/pygbe#57 @kyleniemeyer
Error in run_and_plot_lysozome_case.py
When trying to run `performance/run_and_plot_lysozome_case.py`, I get the following error:
```
  File "run_and_plot_lysozome_case.py", line 246, in <module>
    main()
  File "run_and_plot_lysozome_case.py", line 241, in main
    compiled_results = compile_dict_results(files)
  File "run_and_plot_lysozome_case.py", line 47, in compile_dict_results
    compiled_results[k].append(v)
AttributeError: 'numpy.float64' object has no attribute 'append'
```

I can see that `compiled_results` is supposed to be a dict, so I'm not exactly sure what's happening here...

Tag: https://github.com/openjournals/joss-reviews/issues/306
