### 2017-08-17 barbagroup/pygbe#56 @gforsyth
That needs fixing -- if you run the tests from the `tests` directory it should work
