### 2017-08-17 barbagroup/pygbe#56 @kyleniemeyer
Hmm, when I try `pytest tests/` from the top level, I get the same error
