### 2017-08-17 barbagroup/pygbe#56 @ncclementi
I think what Gil tried to say was running them inside the directory by doing `cd tests` and then running `py.test` . In the documentation we explained that, here is the link http://barbagroup.github.io/pygbe/docs/contributing.html
