### 2017-08-17 barbagroup/pygbe#57 @ncclementi
I know what is the problem, It is looking for `results.pickle` in the directory `output` for some reason it's not there. I'll change that and push it.

Besides I'll add docs on extra packages you need to install to run the script.
`pip install client`
`conda install requests`
