### 2017-08-17 c4b8706
Add installation requirements for performance
