### 2017-08-17 e27efa4
Fix problem when appending v
