### 2017-08-21 2bc2d06
Fix tolerance in test for 7 spheres
