### 2017-08-21 349c894
Improve performance script
