### 2017-08-21 8d39961
Fix typo in script
