+++ performance/run_and_plot_lysozome_case.py
@@ -1,0 +1,287 @@
+import os
+import sys
+import numpy
+import clint
+import pickle
+import zipfile
+import requests
+import subprocess
+
+import matplotlib
+matplotlib.use('Agg') #don't use X backend so headless servers don't die
+
+from matplotlib import pyplot
+
+
+def pickleload(filename):
+    """
+    Does what it says.  Loads a pickle (and returns it)
+    """
+    with open(filename, 'rb') as f:
+        results_dict = pickle.load(f, encoding='latin1')
+
+    return results_dict
+
+
+def compile_dict_results(files):
+    """
+    Compiles a dict of lists of like results given a list of
+    dictionaries containing results
+
+    Inputs
+    ------
+        files: list of strings
+            strings of filenames of input dicts
+
+    Returns
+    -------
+        compiled_results: dict of lists
+            dictionary where keyname is same as input dicts but corresponding
+            value is a list containing all results that have the same key
+    """
+    compiled_results = {}
+    for filename in files:
+        results_dict = pickleload(filename)
+        for k, v in results_dict.items():
+            if k in compiled_results.keys():
+                compiled_results[k].append(v)
+            else:
+                compiled_results[k] = [v]
+
+    return compiled_results
+
+
+def richardson_extrapolation(compiled_results):
+    """
+    Performs an estimate of the exact solution using
+    Richardson extrapolation, given by
+
+    f_ex = (f_1 * f_3 - f_2^2) / (f_3 - 2*f_2+f_1)
+
+    where f_1 is a result from the finest grid and f_3 is from the coarsest.
+    The grids f_1, f_2, f_3 should have the same refinement ratio (e.g. 2 -> 4 -> 8)
+    """
+
+    try:
+        esolv = compiled_results['E_solv_kJ']
+    except KeyError:
+        print('No results found for solvation energy.  \n'
+              'Something has gone wrong.')
+        sys.exit()
+    f1 = esolv[5] # assuming 6 runs: 1, 2, 4, 8, 12, 16
+    f2 = esolv[3]
+    f3 = esolv[2]
+
+    return (f1 * f3 - f2**2) / (f3 - 2 * f2 + f1)
+
+
+
+def generate_plot(compiled_results, filetype='pdf'):
+    """
+    Generates a plot with some hard-coded info based on APBS runs
+    """
+    res = compiled_results
+    N_fem = numpy.array([97*65*97, 129*97*129, 161*129*161, 257*161*257, 385*257*385, 449*385*449,513*449*513])
+    Vfem = 50.*40.*50.
+    Lfem = (Vfem/N_fem)**(1/3.)
+    Lfem_aux = numpy.array([[0.521,0.625,0.521],[0.391,0.417,0.391],
+                            [0.312,0.312,0.312],[0.195,0.250,0.195],
+                            [0.130,0.156,0.130],[0.112,0.104,0.112],
+                            [0.098,0.089,0.098]])
+    timeAPB = numpy.array([4.3,8.6,17.7,54,161,352,768])
+    EsolvAPB = numpy.array([-2237,-2172.9,-2142,-2121.5,-2102.6,-2093.7,-2090.7])
+    apb_ext = -2070.47
+
+    pyg_ext = richardson_extrapolation(res)
+
+    font = {'family':'serif', 'size':7}
+    pyplot.figure(figsize=(3, 2), dpi=80)
+    pyplot.rc('font', **font)
+
+    #calc plot extremas for plotting
+    xmax = N_fem[-1]*5
+    xmin = res['total_elements'][0] / 1.5
+
+    pyplot.semilogx(res['total_elements'], res['E_solv_kJ'],
[…]
