### 2017-08-21 barbagroup/pygbe#57 @ncclementi
@kyleniemeyer I fixed the script and files that were not in the proper format. Now you should be able to reproduce the figure or run your own performance tests.
