### 2017-09-07 barbagroup/pygbe#56 @kyleniemeyer
@ncclementi @gforsyth OK thanks! I pulled the latest version, and everything passes when running `py.test` or `pytest` inside the `tests` directory.

However, I believe `pytest` is the preferred entry point now (https://github.com/pytest-dev/pytest/issues/1629), so you may want to update the documentation to reflect that.
