### 2017-09-07 barbagroup/pygbe#58 @kyleniemeyer
Minor bug on Contributing page in documentation
There's a typo in the Generating Documentation section of the Contributing page in the docs (http://barbagroup.github.io/pygbe/docs/contributing.html#generating-documentation).

Tag: openjournals/joss-reviews#306
