### 2017-09-07 barbagroup/pygbe#57 @kyleniemeyer
OK thanks, with the latest version this now runs ok!

A few comments:

- The link to the notebook on http://barbagroup.github.io/pygbe/docs/run.html#performance seems to be broken
- The figures generated don't 100% match what's in the notebook, with some weird line—looks like a plotting issue:
[time_lys.pdf](https://github.com/barbagroup/pygbe/files/1283525/time_lys.pdf)
[Esolv_lys.pdf](https://github.com/barbagroup/pygbe/files/1283526/Esolv_lys.pdf)
