### 2017-09-07 993ca35
Formatting fixes to contributing.rst
Addresses #58, and also makes the list of contributing rules easier to read
