### 2017-09-07 #306 @kyleniemeyer
Regarding the community guidelines, there is a short section in the Developer's Guide docs page (http://barbagroup.github.io/pygbe/docs/contributing.html) that discusses contributing.

This is probably sufficient, but putting a link to this or a "Help" section in the README (since most people will likely enter the project via it on the main repo page) might strengthen this area.

Also, you might consider adding a code of conduct to the project, and mentioning that in both places.
