### 2017-09-07 #306 @kyleniemeyer
The installation instructions are great—the only hiccup I faced was building PyCUDA on one system. You might consider checking for compatibility/updating with the latest version, PyCUDA 2017.1.1.
