### 2017-09-07 #306 @kyleniemeyer
Regarding the article, I have a number of comments and suggestions:

- I think you should add more at the beginning that explains the target problems in more general terms. In particular, each of "biomolecular electrostatics", " nanoparticle plasmonics ", and "boundary integral method" could be briefly explained. Similarly, LSPR isn't really explained—overall, the third paragraph is nearly impossible to understand (for me, at least) due to the heavy use of jargon.

  (I believe our changes to expectations/guidelines for article content and length happened after you originally submitted this.)

- Also, personally I recommend that references not be treated as objects in the text (i.e., rather than "see @Mayergoyz2007", do "see Mayergoyz and Zhang (@Mayergoyz2007)", but I recognize this is a style preference. Just a suggestion :smile:

- For the last sentence, I would at minimum change to "We plan to use PyGBe-LSPR to study nanobiosensors and to explore nanophotonics applications." (originally missing period at the end, and "to the study of" -> "to study")

- I see reference @CooperClementiBarba2015 in the .bib file; you might consider adding a mention of that to the paper, if the software was used for that study (but update the reference info).
