### 2017-09-07 #306 @kyleniemeyer
Regarding running the code itself, my main comment is that the [Run PyGBe](http://barbagroup.github.io/pygbe/docs/run.html) page could include more information about what is output to the screen when running.

Perhaps this is due to me not being familiar with the problems being solved, but explanations of `Esolv`, `Esurf`, `Ecoul`, or `Time Cext` and `Cext per surface` would help there.
