### 2017-09-07 c11b768
Fixed major PEP8 violation (long line) in main.py
tag https://github.com/openjournals/joss-reviews/issues/306
