### 2017-09-07 d508130
Update docs after building Travis build 226 of barbagroup/pygbe
The docs were built from the branch 'master' against the commit
c580bd9f72f0e08fce2958d6854ab5548325cd88.

The Travis build that generated this commit is at
https://travis-ci.org/barbagroup/pygbe/jobs/272896760.

The doctr command that was run is

    /home/travis/miniconda/envs/pygbe/lib/python3.5/site-packages/doctr/__main__.py deploy --sync --no-require-master --built-docs docs/_build/html docs-master
