### 2017-09-07 barbagroup/pygbe#57 @ncclementi
@kyleniemeyer about:

1. I fixed the broken link, there was an underscore missing. now works
2. I've just run the script again in two different machines (linux and mac), and we don't have the plotting problem in this end. The pdf's on this side look good:
[time_lys.pdf](https://github.com/barbagroup/pygbe/files/1284923/time_lys.pdf)
[Esolv_lys.pdf](https://github.com/barbagroup/pygbe/files/1284924/Esolv_lys.pdf)

Which version of `matplotlib` are you using? I tested `2.0.0` and `2.0.2` and no problem with the figures.
