### 2017-09-07 13a6707
Add missing (yes/no) in one of the questions
