### 2017-09-07 barbagroup/pygbe#57 @kyleniemeyer
I'm using `matplotlib 2.0.2`.
<details><summary>This is the full list of packages in my conda environment</summary>

```
args                      0.1.0                    py35_0    conda-forge
asn1crypto                0.22.0                   py35_0    conda-forge
blas                      1.1                    openblas    conda-forge
ca-certificates           2017.7.27.1                   0    conda-forge
certifi                   2017.7.27.1              py35_0    conda-forge
cffi                      1.10.0                   py35_0    conda-forge
chardet                   3.0.4                    py35_0    conda-forge
clint                     0.5.1                    py35_0    conda-forge
cryptography              1.9                      py35_0    conda-forge
cycler                    0.10.0                   py35_0    conda-forge
dbus                      1.10.10                       3    conda-forge
decorator                 4.1.2                     <pip>
expat                     2.2.1                         0    conda-forge
fontconfig                2.12.1                        4    conda-forge
freetype                  2.7                           1    conda-forge
gettext                   0.19.7                        1    conda-forge
glib                      2.51.4                        0    conda-forge
gst-plugins-base          1.8.0                         0    conda-forge
gstreamer                 1.8.0                         2    conda-forge
icu                       58.1                          1    conda-forge
idna                      2.5                      py35_0    conda-forge
jpeg                      9b                            0    conda-forge
libffi                    3.2.1                         3    conda-forge
libgfortran               3.0.0                         1
libiconv                  1.14                          4    conda-forge
libpng                    1.6.28                        0    conda-forge
libxcb                    1.12                          1    conda-forge
libxml2                   2.9.4                         4    conda-forge
Mako                      1.0.7                     <pip>
MarkupSafe                1.0                       <pip>
matplotlib                2.0.2                    py35_2    conda-forge
ncurses                   5.9                          10    conda-forge
numpy                     1.13.1          py35_blas_openblas_200  [blas_openblas]  conda-forge
openblas                  0.2.19                        2    conda-forge
openssl                   1.0.2l                        0    conda-forge
pcre                      8.39                          0    conda-forge
pip                       9.0.1                    py35_0    conda-forge
py                        1.4.34                    <pip>
pycparser                 2.18                     py35_0    conda-forge
pycuda                    2017.1                    <pip>
PyGBe                     0.3+1.gb26e3a3            <pip>
PyGBe                     0.3+16.ge9c49cd           <pip>
pyopenssl                 16.2.0                   py35_0    conda-forge
pyparsing                 2.2.0                    py35_0    conda-forge
pyqt                      5.6.0                    py35_4    conda-forge
pysocks                   1.6.7                    py35_0    conda-forge
pytest                    3.2.1                     <pip>
python                    3.5.4                         0    conda-forge
python-dateutil           2.6.1                    py35_0    conda-forge
pytools                   2017.4                    <pip>
pytz                      2017.2                   py35_0    conda-forge
qt                        5.6.2                         3    conda-forge
readline                  6.2                           0    conda-forge
requests                  2.18.3                   py35_0    conda-forge
scipy                     0.19.1          py35_blas_openblas_202  [blas_openblas]  conda-forge
setuptools                36.2.2                   py35_0    conda-forge
sip                       4.18                     py35_1    conda-forge
six                       1.10.0                   py35_1    conda-forge
sqlite                    3.13.0                        1    conda-forge
swig                      3.0.12                        2    conda-forge
tk                        8.5.19                        2    conda-forge
tornado                   4.5.1                    py35_0    conda-forge
urllib3                   1.21.1                   py35_1    conda-forge
wheel                     0.29.0                   py35_0    conda-forge
xorg-libxau               1.0.8                         3    conda-forge
xorg-libxdmcp             1.1.2                         3    conda-forge
xz                        5.2.3                         0    conda-forge
zlib                      1.2.11                        0    conda-forge}}
```
