### 2017-09-07 #306 @kyleniemeyer
If the above issues can be addressed satisfactorily, and the issues I raised in the repo itself addressed/closed, then I would consider the submission ready to accept.
