### 2017-09-07 8b79255
Fix overlapping in plotting
