### 2017-09-07 barbagroup/pygbe#57 @ncclementi
Ok, I think I know what is going on, If I'm not wrong:

After you say yes to reproduce the figure in the notebook, you also said yes to run with your local hardware, and since the name of the figure file is the same, we have those weird lines. It's grabbing all the results and plotting and since your local version comes after the results generated to reproduce the figure, it's joining the last point from the repro results to the first point of yours.

If this is the case I've just a pushed a fix, that plots the results we got separately from the once that the user gets.
