### 2017-09-07 barbagroup/pygbe#57 @kyleniemeyer
That does appear to be the issue, but the latest version didn't fix the problem :/
