### 2017-09-07 barbagroup/pygbe#57 @ncclementi
How many runs do you have in the output folder?

When you run the script, there is a point that ask you:
```
There are already results in your output directory.  Do you want to re-run the tests?  If you select "no"
then the plotting routine will still run.  If you select "yes", note that this script is not smart enough to distinguish between "old" and "new" runs and will just jam everything together into one figure (yes/no):
```
