### 2017-09-07 barbagroup/pygbe#57 @kyleniemeyer
@ncclementi Ah, yep that was the issue. After I removed everything from `output` and ran again, it generated the clean figures. Guess I should have read the prompts more carefully...
