### 2017-09-07 #306 @ncclementi
Thanks @kyleniemeyer
We're are working on the suggestions, we'll address them and get back to you here.
