### 2017-09-21 62f3cad
Add docs in dockerfile
