### 2017-09-21 b42deb1
Add docker instructions and update versiosn
