### 2017-09-21 899847d
Update docs after building Travis build 232 of barbagroup/pygbe
The docs were built from the branch 'try_docker' against the commit
b42deb1739673043fe2b4138854638921aea8576.

The Travis build that generated this commit is at
https://travis-ci.org/barbagroup/pygbe/jobs/278288820.

The doctr command that was run is

    /home/travis/miniconda/envs/pygbe/lib/python3.5/site-packages/doctr/__main__.py deploy --sync --no-require-master --built-docs docs/_build/html docs-try_docker
