### 2017-10-17 f8d5438
Reorganize readme
