### 2017-10-17 1e64aae
Add link to contribution guide in README
