### 2017-10-17 ea38502
Update docs after building Travis build 237 of barbagroup/pygbe
The docs were built from the branch 'develop' against the commit
6810d7d822a2f8e5be04c1f251179646e24a403c.

The Travis build that generated this commit is at
https://travis-ci.org/barbagroup/pygbe/jobs/289105813.

The doctr command that was run is

    /home/travis/miniconda/envs/pygbe/lib/python3.5/site-packages/doctr/__main__.py deploy --sync --no-require-master --built-docs docs/_build/html docs-develop
