### 2017-10-17 60def1e
Fix print statements
