### 2017-10-17 748fa73
Fix print statements
