### 2017-10-17 93f1092
Add enter line in printing statements
