### 2017-10-17 5a81f0d
edit first paragraph: mention previous code version
