### 2017-10-17 fe70716
rewrite third paragraph
