+++ paper.md
@@ -30,9 +30,9 @@ biomolecular electrostatics and nanoparticle plasmonics.
 The previous code release, reported in @DCooper2016, solves the Poisson-Boltzmann equation
 for biomolecules immersed in an ionic solvent, using the boundary integral method.
 It computes the solvation energy, which is the free energy spent in moving a biomolecule
-from vaccum to its dissolved state.
-This quantity is used for assessing binding affinity, protein-surface interactions, 
-and other mechanisms at this scale.
+from vacuum to its dissolved state.
+This quantity is used for assessing binding affinity, protein-surface interactions 
+(@CooperClementiBarba2015), and other mechanisms at this scale.
 
 This PyGBe release makes the following contributions:
 (1) it updates the exisiting library presented in @DCooper2016 to Python 3,
@@ -40,19 +40,27 @@ This PyGBe release makes the following contributions:
 (3) it includes better regression tests using pytest and a redesign of the convergence tests.
 
 
-The largest contribution is extending PyGBe to nanoplasmonics, 
-by treating localized surface plasmon resonance (LSPR)
-quasi-statically (see @Mayergoyz2007). LSPR is an optical
-effect (see @Bohren1983), but electrostatics is a good approximation in the long-wavelength
-limit. We use an integral formulation (see @Jung2010), making the existing boundary
-approach suitable and able to exploit the algorithmic and hardware 
-accelaration detailed in @CooperBardhanBarba2014, via the [Barnes-Hut](https://en.wikipedia.org/wiki/Barnes–Hut_simulation) treecode (@BarnesHut1986).
+The largest contribution in this release is extending PyGBe to nanoplasmonics, 
+by treating localized surface plasmon resonance (LSPR) quasi-statically (see @Mayergoyz2007).
+LSPR is essentially a miniaturization of SPR: the resonance of the electron cloud on a 
+metallic surface, excited by incident light.
+It is an optical effect (see @Bohren1983), but electrostatics is a good approximation in the 
+long-wavelength limit. This leads to a coupled system of Poisson equations on complex dielectric regions.
+We use an integral formulation (see @Jung2010), making the existing boundary integral
+approach suitable. 
+The code exploits algorithmic speedup via the Barnes-Hut treecode (@BarnesHut1986), 
+as detailed in @CooperBardhanBarba2014.
+The complex scenario required adapting the linear solver (a GMRES algorithm), 
+modifying the right-hand side, and being able to use the existing treecode
+separately on the real and imaginary parts of the resulting system.
+
+PyGBe's LSPR computations measure the scattered electromagnetic field on a detector 
+that is located far away from a nanoparticle.
 For nanoparticles smaller than the wavelength of incident light, PyGBe 
 can compute the extinction cross-section of absorbing and non-absorbing media
 @Mishchenko2007. 
 
-
-To our knowledge, PyGBe is the only software that uses a fast algorithm—O(N logN),
+To our knowledge, PyGBe is the only open-source software that uses a fast algorithm—O(N logN),
 for N unknowns—and hardware acceleration on GPUs to compute the extinction cross-sections 
 of arbitrary geometries. We plan to use PyGBe-LSPR to the study of nanobiosensors and to explore
 nanophotonics applications
