### 2017-10-17 1577dc8
Update docs after building Travis build 242 of barbagroup/pygbe
The docs were built from the branch 'develop' against the commit
82f67f0ffbb8a60f4d65e8e77365800c648d81d6.

The Travis build that generated this commit is at
https://travis-ci.org/barbagroup/pygbe/jobs/289251038.

The doctr command that was run is

    /home/travis/miniconda/envs/pygbe/lib/python3.5/site-packages/doctr/__main__.py deploy --sync --no-require-master --built-docs docs/_build/html docs-develop
