### 2017-10-17 7f737b6
Update docs after building Travis build 243 of barbagroup/pygbe
The docs were built from the branch 'master' against the commit
6326974ce0676baea5dd9c844277a5955a34c653.

The Travis build that generated this commit is at
https://travis-ci.org/barbagroup/pygbe/jobs/289252400.

The doctr command that was run is

    /home/travis/miniconda/envs/pygbe/lib/python3.5/site-packages/doctr/__main__.py deploy --sync --no-require-master --built-docs docs/_build/html docs-master
