### 2017-10-17 #306 @labarba
Responses to reviewer comments in order of appearance:

> Regarding the community guidelines, there is a short section in the Developer's Guide docs page (http://barbagroup.github.io/pygbe/docs/contributing.html) that discusses contributing.

> This is probably sufficient, but putting a link to this or a "Help" section in the README (since most people will likely enter the project via it on the main repo page) might strengthen this area.

We added a sentence in the README, with a link to the Developoer's Guide.

>Also, you might consider adding a code of conduct to the project, and mentioning that in both places.

We thought about this, and decided that for now, this would be overkill, because we haven't had any contributors from outside the group. We'll revisit this in the future, as needed.

> The installation instructions are great—the only hiccup I faced was building PyCUDA on one system. You might consider checking for compatibility/updating with the latest version, PyCUDA 2017.1.1.

We updated the compatibility of dependencies for the latest versions of all our packages, and we added a Docker file to make the installation process easier (in case the user is familiarized with Docker).

> I think you should add more at the beginning that explains the target problems in more general terms. In particular, each of "biomolecular electrostatics", " nanoparticle plasmonics ", and "boundary integral method" could be briefly explained. Similarly, LSPR isn't really explained—overall, the third paragraph is nearly impossible to understand (for me, at least) due to the heavy use of jargon.

In the revised first paragraph, we explain "biomolecular electrostatics" by saying what we compute: the solvation energy, and explaining what this is (and what it is used for). We don't explain "boundary integral method," because of the same reason we wouldn't expalin "finite difference method"—it is a standard method in the arsenal of numerical computing. We revised the third paragraph, explaining LSPR, which also helps get an idea of what part of nanoplasmonics we're tackling.

> Also, personally I recommend that references not be treated as objects in the text (i.e., rather than "see @mayergoyz2007", do "see Mayergoyz and Zhang (@mayergoyz2007)", but I recognize this is a style preference. Just a suggestion.

We agree completely. But notice that the JOSS machinery produces the correct format when generating the PDF, with Author (YEAR) spelled out, as needed when part of a sentence. In other words, the way we have it in the `paper.md` file will produce what you want in the final PDF.

> For the last sentence, I would at minimum change to "We plan to use PyGBe-LSPR to study nanobiosensors and to explore nanophotonics applications." (originally missing period at the end, and "to the study of" -> "to study")

Fixed.

> I see reference @cooperclementibarba2015 in the .bib file; you might consider adding a mention of that to the paper, if the software was used for that study (but update the reference info). Also, maybe the other papers that are mentioned in the docs?

That paper used the previous version of the code, and we added the reference in the first paragraph.

> Regarding running the code itself, my main comment is that the Run PyGBe page could include more information about what is output to the screen when running.

> Perhaps this is due to me not being familiar with the problems being solved, but explanations of Esolv, Esurf, Ecoul, or Time Cext and Cext per surface would help there.

We modified the printed output so that a short description appears before each instance of the returned variables. Morover, there is a section in the main page of the Documentation (now renamed "Code Documentation:") where a user can search by variable name or keyword, to be pointed to the relevant docstring(s).
