### 2017-10-17 #306 @kyleniemeyer
Thanks @labarba, looks good!

@katyhuff at this point I recommend accepting the submission.
