### 2017-10-18 7935e5e
Fix variable name to avoid casting error
