### 2017-10-19 #306 @katyhuff
Wonderful! Thanks @labarba @ncclementi and @kyleniemeyer !

@labarba and @ncclementi : Could you make an archive of the current version of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
