### 2017-11-17 220ed70
Update docs after building Travis build 248 of barbagroup/pygbe
The docs were built from the branch 'tree_bug' against the commit
1ccf168a417025c0bd6e271b34d01dc0cd31d301.

The Travis build that generated this commit is at
https://travis-ci.org/barbagroup/pygbe/jobs/303596940.

The doctr command that was run is

    /home/travis/miniconda/envs/pygbe/lib/python3.5/site-packages/doctr/__main__.py deploy --sync --no-require-master --built-docs docs/_build/html docs-tree_bug
