+++ docs-tree_bug/_sources/source/modules.rst.txt
@@ -1,0 +1,7 @@
+pygbe
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   pygbe
