+++ docs-tree_bug/_sources/index.rst.txt
@@ -1,0 +1,25 @@
+.. include:: intro.rst
+
+.. toctree::
+   :titlesonly:
+   :maxdepth: 2
+
+   installation
+   run
+   cite
+   contributing
+   mesh_pqr_setup.rst
+   input_format
+   references
+   othersoftware
+   programmatic
+
+
+
+Code Documentation
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
+
