### 2017-11-20 #306 @arfon
> @labarba and @ncclementi : Could you make an archive of the current version of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.

Friendly reminder on this @ncclementi.
