### 2017-11-20 #306 @labarba
This is me being a bottleneck. We had the problem of the Zenodo archive being created by one student who has now left the group. Zenodo does not have the feature to allow org accounts or sharing the ownership of a Zenodo entry. We exchanged emails with them to have this archive transferred to another account. They said yes, with approval of original uploader. This happened. Now, I have to create a new Zenodo account with a "barbagroup" org email I created, and transfer ownership of the Zenodo entry. Will update here when done.
