### 2017-11-21 ae12f95
Modify Dockerfile to install from master branch
