### 2017-11-21 4944d99
Fix errors in Dockerfile
