### 2017-11-28 #306 @ncclementi
@arfon Here is our zenodo doi: https://doi.org/10.5281/zenodo.1067773
