### 2017-11-30 1cd7b58
Updating paper YAML header
