### 2017-11-30 #306 @arfon
@ncclementi @labarba - when you get a chance, please merge [this PR](https://github.com/barbagroup/pygbe/pull/66) which updates the JOSS paper to the newer format.
