### 2017-11-30 #306 @arfon
@kyleniemeyer many thanks for your review and to @katyhuff for editing this submission ✨

@ncclementi @labarba - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00306 ⚡️ 🚀 💥
